test(mocker): trim low-signal replay coverage
